Rename send_audio.py to audio_transcription_client.py
- Rename file from send_audio.py to audio_transcription_client.py to match class name
- Update all import statements to use the new filename:
  * src/core/application.py: Update AudioTranscriptionClient import
  * src/clients/__init__.py: Update send_audio_file import
- Update usage examples in the file to reference new filename
- Update README.md to reflect the new filename in project structure
- All 66 unit tests pass after the rename

Benefits:
- Better naming consistency between file and class names
- Improved code organization and readability
- Follows Python naming conventions
- Maintains all existing functionality
