+++ src/clients/__init__.py
@@ -4,7 +4,7 @@ Client implementations
 Contains client code for different services
 """
 
-from .send_audio import send_audio_file
+from .audio_transcription_client import send_audio_file
 from .infer_client import transcribe
 
 __all__ = ['send_audio_file', 'transcribe']
